Dewan-Mehra/requirements.txt
streamlit
google-generativeai
requests
matplotlib